requirements.txt
# Dependensi Python untuk CCTV AI Telegram Bot
# Dikhususkan untuk Ubuntu 22.04

# OpenCV untuk pemrosesan gambar dan video
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# Pengenalan wajah dan deteksi
face-recognition==1.3.0
numpy==1.24.3

# Telegram Bot
python-telegram-bot==20.7

# Pengolahan gambar
Pillow==10.0.1

# HTTP requests untuk kamera IP
requests==2.31.0

# Konfigurasi YAML
pyyaml==6.0.1

# Komputasi ilmiah
scipy==1.11.3

# Library untuk dlib (dibutuhkan face-recognition)
dlib==19.24.2

# Runtime untuk model ONNX
onnxruntime==1.16.3

# YOLOv8 untuk deteksi orang yang lebih akurat
ultralytics==8.0.196
cvzone==1.6.1

# Logging
colorlog==6.8.0

# Async support
aiohttp==3.9.1

# Time handling
python-dateutil==2.8.2

# Path handling
pathlib==1.0.1
